+++ ItemQualities/Assets/ItemQualities/Scripts/CharacterMasterExtraStatsTracker.cs
@@ -40,7 +40,7 @@ namespace ItemQualities
         public int BossDamageBonusTicks;
 
         [SyncVar]
-        public int CardStoredInteractableIndex = -1;
+        public StoredInteractableInfo CardStoredInteractableInfo = StoredInteractableInfo.None;
 
         readonly SyncListUInt _upgradeItemIndices = new SyncListUInt();
 
@@ -427,7 +427,10 @@ namespace ItemQualities
 
         void onIncomingDamageServer(DamageInfo damageInfo)
         {
-            if (damageInfo.damage > 0f && !damageInfo.delayedDamageSecondHalf && (damageInfo.damageType & DamageType.DoT) == 0)
+            if (damageInfo.damage > 0f &&
+                !damageInfo.delayedDamageSecondHalf &&
+                (damageInfo.damageType & DamageType.DoT) == 0 &&
+                !damageInfo.IsParried())
             {
                 _stageIncomingDamageInstanceCountServer++;
                 OnStageDamageInstancesTakenCountChangedServer?.Invoke(this);